Read the int32 bag-result the 3.4.3 inventory failure packet carries
SMSG_INVENTORY_CHANGE_FAILURE (InventoryChangeFailure::Write) leads with
an int32 BagResult; the old uint8 read only worked by accident through
the little-endian low byte because the InventoryResult enum is uint8.
Read the actual wire width so the error switch stays correct if error
codes ever grow past 255.

+++ src/plugins/playerbot/strategy/actions/InventoryChangeFailureAction.cpp
@@ -9,7 +9,10 @@ bool InventoryChangeFailureAction::Execute(Event event)
 {
     WorldPacket p(event.getPacket());
     p.rpos(0);
-    uint8 err;
+    // SMSG_INVENTORY_CHANGE_FAILURE leads with int32 BagResult
+    // (InventoryChangeFailure::Write); the InventoryResult enum is uint8, so a
+    // uint8 read only worked by accident via the little-endian low byte.
+    int32 err;
     p >> err;
     if (err == EQUIP_ERR_OK)
         return false;
